[fix] Fix parser ambiguity: don't allow juxtaposition in embeddings
Now, when embedding `Syn` nouns in a `Chk` juxtaposition, we always
prefer `(Emb a, Emb b)`, to `Emb (a, b)`. This stops the parser
assuming that juxtapositions are `Syn` based on their first argument
and should hence fix the bug where applications can't have a variable
followed by a number literal.