refactor: Combine hugr types and ops into one module (#51)
Once we upgrade hugr there's a lot of mutual dependency in the data
structures, like HugrValue -> TypeArg -> HugrType -> NodeOp -> Hugr ->
HugrValue. Sadly this means it becomes a pain to mimick the file
structure used in guppy with a separation between ops and types. Guppy
can do it because in python you can use types that haven't been defined